pyproject.toml
[tool.poetry]
name = "siren"
version = "0.1.0"
description = "Scrapers for the SIREN project"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
bs4 = "^0.0.1"
pandas = "^2.0.1"
selenium = "^4.9.1"
pytesseract = "^0.3.10"
pdf2img = "^0.1.2"
lxml = "^4.9.2"
opencv-contrib-python = "^4.8.0.74"
numpy = "^1.25.1"
opencv-python = "^4.8.0.74"
pillow = "^10.0.0"
google-auth = "^2.22.0"
google-auth-oauthlib = "^1.0.0"
google-auth-httplib2 = "^0.1.0"
google-api-python-client = "^2.94.0"
yarl = "^1.9.2"
black = "^23.7.0"
isort = "^5.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
